fundamental error
validators are run on property values, or in the case of a property
array, each item in that array.

This means validators are never run on arrays as a whole. We can’t, for
example, validate array length.

It also destroys the required validator.

The type sanitizer sanitizes null to null, which causes the required
validator to fail, as expected.

However, the type sanitizer also sanitizes [null] to [], which makes
the required validator pass.

This means that, without sanitization, [null] would fail and [] would
pass.

Unacceptable.

I need to come up with an elegant way to specify validators which are
supposed to run on array properties, array items, or both.