fixed fundamental problem
okay, so validators and sanitizers don’t just run in item values, they
have to make considerations in case their property is an array property.

And I’ve come across some problems with my validator config parser, but
i’ll wait till everything else is done to worry about that.

As it stands, every validator and sanitizer now accounts for array
properties. I’d like to do some more testing with larger schemas, but
after that I’m happy to call it a beta, write the documentation and
release it to npm